refactor: grid components
